to reduce the load of CI move fault controllable fifo to tester subproject.
